pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "domain-synth-auditor"
version = "0.1.0"
description = "PoC агентной системы для аудита синтетических изображений в доменно-специфичных сценариях"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "aiohttp>=3.9",
    "aiomysql>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
